Add detail month

+++ presentation_common/src/main/java/com/ssepulveda/presentation_common/inputs/MonthInput.kt
@@ -1,4 +1,4 @@
 package com.ssepulveda.presentation_common.inputs
 
 
-data class MonthInput(val monthId: Long)
+data class MonthInput(val monthId: Int)
